requirements.txt
requests~=2.27.1
python-mpv-jsonipc
python-mpv
appdirs~=1.4.4
getch